add TrackQueue and client credentials token flow
Within the `TrackQueue` is a flow where a timer is started for the
duration of the song. After the timer gets to 0, it will start a new
timer for the next track and remove the current one. The goal of which
is to replicate how a radio works, giving more of a live radio
experience.