there is a lot going on here some of which is fluff
There was some experimentation going on to figure out a way to do
refresh logic for Spotify. Once I discovered that you can't use a
refresh token to obtain _another_ refresh token for a future refresh, I
went back to the drawing board for how to handle it. A lot of this was
setting up the plumbing to make that work.

What I am currently working to do is keep a Context representing a user
and their choices for refreshing (page or token). If an API call becomes
unauthorized, then I redirect to /login and it should bring them back to
the radio station to retry the call. Not entirely ideal, but it works.