implement logout
This essentially deletes the user's context from the `UserContext`
GenServer. If that user then proceeds to visit /radio/anything they will
become logged in again. This will make the app proceed to seem like the
logout did not work when in reality it did. It will seem that way
because /login does magic with Spotify where it will log the user in at
that point.

To prevent that entirely where they are completely signed out of Spotify
is something I need to investigate further. I'm assuming Spotify has a
cookie or something that is being leveraged to bypass giving the app
access via username/password.